Refactor of driver.ts
